Formats

StructPack.Format — Type

Abstract format type.

Formats determine the rules for packing and unpacking values via msgpack primitives. They are supposed to be singleton types.

To add support for a new format F <: Format, define the corresponding methods of pack and unpack.

This package comes with a number of built-in formats. The following core formats have low-level implementations that build upon one or more formats of the msgpack specification:

NilFormat (msgpack nil),

BoolFormat (msgpack boolean),

SignedFormat (msgpack negative / positive fixint, signed 8-64),

UnsignedFormat (msgpack positive fixint, unsigned 8-64),

FloatFormat (msgpack float32, float64)

StringFormat (msgpack fixstr, str 8-32),

BinaryFormat (msgpack bin 16, bin 32).

For vector-like and map-like objects, several built-in formats with different benefits and drawbacks are provided as subtypes of

AbstractVectorFormat (msgpack fixarray, array 16, array 32),

AbstractMapFormat (msgpack fixmap, map 16, map 32).

Additional convenience formats include

ArrayFormat (store multidimensional arrays),

BinVectorFormat (store vectors with bitstype elements efficiently),

BinArrayFormat (store multidimensional bitstype arrays efficiently),

TypeFormat (store types)

TypedFormat (store values and their type for generic unpacking).

Basic formats

StructPack.NilFormat — Type

Core format for packing nil values.

Built upon the msgpack format nil.

Defaults

NilFormat is the default format of Nothing. Use

format(::Type{T}) = NilFormat()

or

@pack T in NilFormat

to make NilFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

All types can be packed in NilFormat.

Unpacking

To support unpacking values of type T packed in NilFormat, implement

construct(::Type{T}, ::Nothing, ::NilFormat)::T

or make sure that the constructor T() is defined.

StructPack.BoolFormat — Type

Core format for packing boolean values.

Built upon the msgpack format boolean.

Defaults

BoolFormat is the default format of Bool. Use

format(::Type{T}) = BoolFormat()

or

@pack T in BoolFormat

to make BoolFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in BoolFormat, implement

destruct(val::T, ::BoolFormat)::Bool

Unpacking

To support unpacking values of type T packed in BoolFormat, implement

construct(::Type{T}, ::Bool, ::BoolFormat)::T

or make sure that the constructor T(::Bool) is defined.

StructPack.SignedFormat — Type

Core format for packing signed integer values.

Built upon the msgpack formats negative fixint, positive fixint, signed 8, signed 16, signed 32, signed 64.

Defaults

SignedFormat is the default format of all subtypes of Signed. Use

format(::Type{T}) = SignedFormat()

or

@pack T in SignedFormat

to make SignedFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in SignedFormat, implement

destruct(val::T, ::SignedFormat)::Signed

Unpacking

To support unpacking values of type T packed in SignedFormat, implement

construct(::Type{T}, ::Int64, ::SignedFormat)::T

or make sure that the constructor T(::Int64) is defined.

StructPack.UnsignedFormat — Type

Core format for packing unsigned integer values.

Built upon the msgpack formats positive fixint, unsigned 8, unsigned 16, unsigned 32, unsigned 64.

Defaults

UnsignedFormat is the default format of all subtypes of Unsigned. Use

format(::Type{T}) = UnsignedFormat()

or

@pack T in UnsignedFormat

to make UnsignedFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in UnsignedFormat, implement

destruct(val::T, ::UnsignedFormat)::Unsigned

Unpacking

To support unpacking values of type T packed in UnsignedFormat, implement

construct(::Type{T}, ::UInt64, ::UnsignedFormat)::T

or make sure that the constructor T(::UInt64) is defined.

StructPack.FloatFormat — Type

Core format for packing float values.

Built upon the msgpack formats float 32, float 64.

Defaults

FloatFormat is the default format for Float16, Float32, and Float64. Use

format(::Type{T}) = FloatFormat()

or

@pack T in FloatFormat

to make FloatFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in FloatFormat, implement

destruct(val::T, ::FloatFormat)::Union{Float16, Float32, Float64}

Unpacking

To support unpacking values of type T packed in FloatFormat, implement

construct(::Type{T}, ::Float64, ::FloatFormat)::T

or make sure that the constructor T(::Float64) is defined.

StructPack.StringFormat — Type

Core format for packing string values.

Built upon the msgpack formats fixstr, str 8, str 16, str 32.

Defaults

StringFormat is the default format for Symbol and all subtypes of AbstractString. Use

format(:: Type{T}) = StringFormat()

or

@pack T in StringFormat

to make StringFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in StringFormat, implement

destruct(val::T, ::StringFormat)::R

where the returned value ret::R must implement sizeof(ret) (number of bytes) as well as write(io, ret).

Unpacking

To support unpacking values of type T packed in StringFormat, implement

construct(:: Type{T}, ::String, ::StringFormat)::T

or make sure that convert(T, ::String) is defined.

StructPack.BinaryFormat — Type

Core format for packing binary values.

Built upon the msgpack formats bin 8, bin 16, bin 32.

Defaults

Use

format(::Type{T}) = BinaryFormat()

or

@pack T in BinaryFormat

to make BinaryFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in BinaryFormat, implement

destruct(val::T, ::BinaryFormat)::R

where the returned value ret::R must implement sizeof(ret) (number of bytes) as well as write(io, ret).

Unpacking

To support unpacking values of type T packed in BinaryFormat, implement

construct(::Type{T}, ::Vector{UInt8}, ::BinaryFormat)::T

or make sure that the constructor T(::Vector{UInt8}) is defined.

StructPack.ExtensionFormat — Type

Format for supporting the msgpack extension standard.

Built upon the msgpack formats fixext 1 - 8 and ext 8 - 32.

Defaults

Use

format(::Type{T}) = ExtensionFormat{I}()

or

@pack T in ExtensionFormat{I}

to make ExtensionFormat with msgpack extension type I::Int8 the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in ExtensionFormat, implement

destruct(val::T, ::ExtensionFormat{I})::R

where the return value ret::R must implement sizeof(ret) (number of bytes) as well as write(io, ret).

Unpacking

To support unpacking values of type T packed in ExtensionFormat, implement

construct(::Type{T}, ::Vector{UInt8}, ::ExtensionFormat{I})::T

or make sure that the constructor T(::Vector{UInt8}) is defined.

Vector and array formats

Besides VectorFormat, which directly mirrors the msgpack vector format, StructPack.jl also provides convenience formats for storing isbits-vectors and multidimensional arrays.

StructPack.AbstractVectorFormat — Type

Umbrella type for VectorFormat and DynamicVectorFormat.

StructPack.VectorFormat — Type

Core format for packing vector values.

Built upon the msgpack formats fixarray, array 16, array 32.

Defaults

VectorFormat is the default format for subtypes of Tuple and AbstractVector. Use

format(::Type{T}) = VectorFormat()

or

@pack T in VectorFormat

to make VectorFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in VectorFormat, implement

destruct(val::T, ::VectorFormat)::R

where the returned value ret::R must implement Base.length(ret) (number of entries) and must be iterable. The formats of the entries of val are determined via valueformat, where state is the linear index.

Unpacking

To support unpacking values of type T packed in VectorFormat, implement

construct(::Type{T}, values::Generator{T}, ::VectorFormat)::T

or make sure that the constructor T(values::Generator{T}) is defined (see Generator). The respective types and formats of the entries of values are determined via valuetype and valueformat, where state is the linear index.

Warning

During construction, all entries of the generator values have to be iterated over. Since Generator{T} wraps a lazy map that reads from the IO source to be unpacked, not conducting the iteration before the next object is processed will interfere with unpacking of subsequent values. For the same reason, you should not store the object values and access it at a later time.

StructPack.ArrayFormat — Type

Format for arrays that remembers the array size.

Built upon VectorFormat.

Defaults

ArrayFormat is the default format for AbstractArray. Use

format(::Type{T}) = ArrayFormat()

or

@pack T in ArrayFormat

to make ArrayFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in ArrayFormat, implement

destruct(val::T, ::ArrayFormat)::R

where the returned value ret::R must define size and must be packable in VectorFormat.

Unpacking

To support unpacking values of type T packed in ArrayFormat, implement

construct(::Type{T}, ::ArrayValue{Generator{T}}, ::ArrayFormat)::T

or make sure that the constructor T(val::ArrayValue{Generator{T}}) is defined (see ArrayValue and Generator).

StructPack.BinVectorFormat — Type

Pack vectors in binary format.

Built upon BinaryFormat.

Defaults

BinVectorFormat is the default format for BitVector. Use

format(::Type{T}) = BinVectorFormat()

or

@pack T in BinVectorFormat

to make BinVectorFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in BinVectorFormat, implement

destruct(val::T, ::BinVectorFormat)::R

where the returned value ret::R must be packable in BinaryFormat.

Unpacking

To support unpacking values of type T packed in BinVectorFormat, implement

construct(::Type{T}, ::Vector{UInt8}, ::BinVectorFormat)::T

or make sure that the constructor T(bytes::Vector{UInt8}) is defined.

StructPack.BinArrayFormat — Type

Pack arrays in binary format.

Built upon BinVectorFormat.

Defaults

BinArrayFormat is the default format for BitArray{N} for N > 1. Use

format(::Type{T}) = BinArrayFormat()

or

@pack T in BinArrayFormat

to make BinVectorFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in BinArrayFormat, implement

destruct(val::T, ::BinArrayFormat)::R

where the returned value ret::R must define Base.size and must be packable in BinaryFormat.

Unpacking

To support unpacking values of type T packed in BinArrayFormat, implement

construct(::Type{T}, ::BinArrayValue{Vector{UInt8}}, ::BinArrayFormat)::T

or make sure that the constructor T(val::BinArrayValue{Vector{UInt8}}) is defined (see [BinArrayValue(@ref)]).

StructPack.DynamicVectorFormat — Type

Modification of VectorFormat.

During unpacking, the types and formats of future entries may depend on past entries via overloading iterstate.

Info

DynamicVectorFormat is currently slower than VectorFormat, even if iterstate just enumerates indices. In principle, however, the compiler should have all information neccessary to optimize DynamicVectorFormat in this case and bring the performance on par. In the future, we might thus deprecate DynamicVectorFormat and absorb its functionality into VectorFormat.

Map formats

Besides MapFormat, which directly mirrors the msgpack map format, StructPack.jl also provides specialized map implementations for structures.

StructPack.AbstractMapFormat — Type

Umbrella type for MapFormat, DynamicMapFormat, and AbstractStructFormat.

StructPack.MapFormat — Type

Core format for packing map / dictionary values.

Built upon the msgpack formats fixmap, map 16, map 32.

Defaults

MapFormat is the default format for subtypes of NamedTuple and Dict. Use

format(::Type{T}) = MapFormat()

or

@pack T in MapFormat

to make MapFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in MapFormat, implement

destruct(val::T, ::MapFormat)::R

where the returned value ret::R must implement Base.length(ret) (number of entries) and must be iterable with key-value pairs as entries. The key / value formats of the entries of val are determined via keyformat and valueformat, where state is the linear index.

Unpacking

To support unpacking values of type T packed in MapFormat, implement

construct(::Type{T}, pairs::Generator{T}, ::MapFormat)::T

or make sure that the constructor T(pairs::Generator{T}) is defined (see Generator), where pairs will contain key-value pairs as entries. The respective types and formats of the keys and values are determined during unpacking via keytype, valuetype, keyformat and valueformat, where state is the linear index.

Warning

During construction, all entries of the generator pairs have to be iterated over. Since Generator{T} wraps a lazy map that reads from the IO source to be unpacked, not conducting the iteration before the next object is processed will interfere with unpacking of subsequent values. For the same reason, you should not store the object pairs and access it at a later time.

StructPack.DynamicMapFormat — Type

Modification of MapFormat.

During unpacking, the types and formats of future entries may depend on past entries via overloading iterstate.

Info

DynamicMapFormat is currently slower than MapFormat, even if iterstate just enumerates indices. In principle, however, the compiler should have all information neccessary to optimize DynamicMapFormat in this case and bring the performance on par. In the future, we might thus deprecate DynamicMapFormat and absorb its functionality into MapFormat.

StructPack.AbstractStructFormat — Type

Umbrella type for StructFormat, UnorderedStructFormat, and FlexibleStructFormat.

StructPack.StructFormat — Type

Format for packing structures.

Built upon the msgpack formats fixmap, map 16, map 32.

Defaults

StructFormat is not used as default for any type. However, it should be the first candidate when you want to pack custom structs. Use

format(::Type{T}) = StructFormat()

or

@pack T in StructFormat

to make StructFormat the default format for type T. If T is abstract, use {<: T} to cover all subtypes.

Packing

To support packing values of type T in StructFormat, implement

destruct(val::T, ::MapFormat)::R

where the returned value ret::R must implement Base.length(ret) (number of entries) and must be iterable with key-value pairs as entries.

In contrast to MapFormat, the keys are always stored in StringFormat and the value formats are determined via fieldformats, which should return a tuple of formats for the value-entries of the struct.

Unpacking

To support unpacking values of type T packed in StructFormat, implement

construct(::Type{T}, pairs::Generator{T}, ::StructFormat)::T

or make sure that the constructor T(values...) is defined, where values contains the value-entries of pairs.

The respective types and formats of the values are determined during unpacking via fieldtypes and fieldformats, which should return tuples of types and formats. Additionally, fieldnames is queried to make sure the keys encountered during unpacking are consistent with T.

StructFormat vs. MapFormat

Both StructFormat and MapFormat can be used to pack custom structs. They perform comparably.

MapFormat is more flexible and can handle keys that are not symbols. In contrast, StructFormat requires static fieldtype, fieldname, and fieldformat information.

However, MapFormat will not counter-check key values during unpacking and can thus easily lead to corrupted data if applied on external msgpack binaries.

StructPack.UnorderedStructFormat — Type

Modification of StructFormat.

This map-based format automatically sorts entries according to fieldnames during unpacking.

While the unpacking performance is deteriorated compared to StructFormat, this format makes it possible to load msgpack binaries where the order of map-entries cannot be guaranteed.

By default, the constructor T(values...) is used when unpacking a type T in UnorderedStructFormat, where values denotes the value-entries unpacked from the msgpack map. To use a keyword-argument based constructor, simply define

construct(::Type{T}, pairs, ::UnorderedStructFormat) = T(; pairs...)

StructPack.FlexibleStructFormat — Type

Modification of StructFormat.

When unpacking a value in map format, this format ignores superfluous fields and supports reading only a subset of the fields of a struct.

By default, the keyword argument based constructor T(; pairs...) is used when unpacking a type T in FlexibleStructFormat, where pairs denotes the entries unpacked from the msgpack map.

Type formats

StructPack.TypeFormat — Type

Format that is used for packing types.

In order to pack and unpack a type T::Type in TypeFormat, make sure that t = StructPack.TypeValue(T) works and that StructPack.composetype(t) successfully reconstructs T.

If T has type parameters, their serialization can be influenced via the functions typeparamtypes and typeparamformats.

By default, all type parameters are packed / unpacked in TypedFormat.

StructPack.TypedFormat — Type

Wrapper format for storing the value and type of an object.

If a value val::T can be packed in the format F<:Format and its type T can be packed in TypeFormat, then packing val in TypedFormat{F} enables unpacking via unpack(io, TypedFormat{F}()), i.e., without knowledge of T.

Special formats

StructPack.DefaultFormat — Type

Special format that serves as lazy placeholder for format(T) in situations where the type T is not yet known.

Warning

Never define format(T) for a type T in terms of DefaultFormat. This will lead to indefinite recursion.

StructPack.AnyFormat — Type

Convenience format to unpack generic msgpack values.

Note that custom contexts (see Context) are ignored by AnyFormat.

Note

The msgpack timestamp format is currently not supported by AnyFormat.

StructPack.SetContextFormat — Type

Special format that overrides the current context.

In particular, packing (or similarly unpacking) via

pack(val, SetContextFormat{C, F}(), C2())

is equivalent to

pack(val, F(), C())

where the original context C2 <: Context is ignored.

This format is for example useful in combination with fieldformats, to enforce that different fields of a struct can be packed / unpacked with different contexts.