API Reference

Main API

StructPack.format — Function

format(T::Type [, ctx::Context])::Format
format(::T [, ctx::Context])::Format

Return the format associated to type T under ctx.

The context-free version of this method must be implemented in order for pack(io, value::T) and unpack(io, T) to work. It is used as fallback for all contexts.

See also Format and DefaultFormat.

StructPack.pack — Function

pack(value, [, ctx::Context])::Vector{UInt8}
pack(value, [, fmt::Format, ctx::Context])::Vector{UInt8}
pack(io::IO, args...)::Nothing

Create a binary msgpack representation of value according to the given format fmt. If a stream io is passed, the representation is written into it.

If no format is provided, it is derived from the type of value via format(typeof(value), ctx). The context defaults to the value of context, which is a scoped value in julia 1.11 or newer and a reference else.

If both a format and a context is provided, fmt is used for packing value while ctx is passed along to deeper packing related calls.

StructPack.unpack — Function

unpack(bytes::Vector{UInt8}, T::Type [, ctx::Context])::T
unpack(bytes::Vector{UInt8}, T::Type [, fmt::Format, ctx:::Context])::T
unpack(io::IO, T::Type, args...)::T

Unpack a binary msgpack representation of a value of type T in format fmt from a byte vector bytes or a stream io. The returned value is guaranteed to be of type T.

If no format is provided, it is derived from T via format(T, ctx). The context ctx defaults to context.

unpack(bytes::Vector{UInt8})::Any
unpack(io::IO)::Any

Unpack a binary msgpack value via the special format AnyFormat. The returned value can be of any type.

StructPack.construct — Function

construct(T::Type, val, fmt::Format [, ctx::Context])::T

Postprocess a value val unpacked according to fmt and return an object of type T. The type of val depends on the format fmt that was used for unpacking.

Defaults to T(val) but can be overwritten for any combination of T, fmt, and ctx.

StructPack.destruct — Function

destruct(val::T, fmt::Format [, ctx::Context])

Preprocess a value val to prepare packing it in the format fmt.

Defaults to val but can be overwritten for any combination of T, fmt, and ctx.

Each format has specific requirements regarding the output of this method.

Keys and values

StructPack.keytype — Function

keytype(T::Type, state, fmt::Format [, ctx::Context])::Type

Return the type of the key at iteration state state when saving the entries of T in format fmt.

This method is called when unpacking values in AbstractMapFormat.

StructPack.keyformat — Function

keyformat(T::Type, state, fmt::Format [, ctx::Context])::Format

Return the format of the key at iteration state state when saving the entries of T in format fmt.

This method is called when packing or unpacking values in AbstractMapFormat.

StructPack.valuetype — Function

valuetype(T::Type, fmt::Format, state [, ctx::Context])::Type

Return the type of the value at iteration state state when saving the entries of T in format fmt.

This method is used when unpacking values in AbstractVectorFormat and AbstractMapFormat.

StructPack.valueformat — Function

valueformat(T::Type, state, fmt::Format [, ctx::Context])::Format

Return the format of the value at iteration state state when saving the entries of T in format fmt.

This method is used when packing or unpacking values in AbstractVectorFormat and AbstractMapFormat.

StructPack.iterstate — Function

iterstate(T::Type, fmt::Format [, ctx::Context])

Initialize a state object that is repeatedly updated while iterating over the entries of values of type T when packing and unpacking in format fmt.

By default, the initial state is 1 and gets replaced by state + 1 in subsequent updates.

This function is only relevant for formats that dynamically make use of the methods keytype, keyformat, valuetype, or valueformat, like DynamicVectorFormat and DynamicMapFormat.

iterstate(T::Type, state, entry, fmt::Format [, ctx::Context])

Return an update of the state object state when T is packed or unpacked in the format fmt.

The argument entry signifies the entry packed / unpacked in the last iteration and can be used to inform the next iteration state. It will be of type valuetype(T, fmt, state, ctx) in case of DynamicVectorFormat and similarly a key-value pair with types determined by keytype and valuetype in case of DynamicMapFormat.

This approach enables 'dynamic' unpacking where the type / format of an entry depends on the values unpacked previously. The format TypedFormat exploits this pattern.

Struct fields

StructPack.fieldformats — Function

fieldformats(T::Type [, ctx::Context])

Return the field types of T when packing / unpacking in AbstractStructFormat under ctx.

StructPack.fieldtypes — Function

fieldtypes(T::Type [, ctx::Context])

Return the field types of T when packing / unpacking in AbstractStructFormat under ctx.

Defaults to Base.fieldtypes(T).

StructPack.fieldnames — Function

fieldnames(T::Type [, ctx::Context])

Return the field names of T when packing / unpacking in AbstractStructFormat under ctx.

Defaults to Base.fieldnames(T).

Auxiliary types

StructPack.Generator — Type

Auxiliary object for unpacking sequential values.

Implements the Base iterator interface and can for most purposes be treated like a Base.Generator. Compared to the latter, it is equipped with additional element type information via its first type parameter T.

In particular, when unpacking a value of type Generator{T} in VectorFormat or MapFormat, the methods valuetype, valueformat, keytype, and keyformat are called with type argument T to determine how elements are unpacked.

StructPack.ArrayValue — Type

Wrapper struct employed when unpacking a value in ArrayFormat.

Contains the fields size and data, the latter of which is usually a Generator after reconstruction via unpack.

StructPack.BinArrayValue — Type

Wrapper object for unpacking values in BinArrayFormat.

Contains the fields size and data, the latter of which is a Vector{UInt8} after reconstruction via unpack.

StructPack.TypeValue — Type

Auxiliary structure that expresses a type in a format friendly to serialization.

StructPack.TypedValue — Type

Auxiliary structure that expresses a value and its type in a format friendly to serialization.

StructPack.ExtensionData — Type

Structure holding the type and binary data stored in a msgpack extension object.

Has the fields type::Int8 and data::Vector{UInt8}.

Typed format

StructPack.typeparamtypes — Function

typeparamtypes(T::Type [, ctx::Context])

Return the types of the type parameters of T when packing / unpacking under ctx.

Note

If T has type parameters, this method must be implemented for packing / unpacking types via TypeFormat and their values via TypedFormat.

StructPack.typeparamformats — Function

typeparamformats(T::Type [, ctx::Context])

Return the formats of the type parameters of T when packing / unpacking under ctx.

Defaults to DefaultFormat() for each type parameter.

This method is consulted when packing / unpacking types via TypeFormat and TypedFormat.

Scoped values

StructPack.context — Constant

Scoped value that captures the active context.

To pack / unpack in a given context ctx::Context, you can use this pattern:

using Base.ScopedValues

with(StructPack.context => ctx) do
  # Do your packing / unpacking without passing ctx explicitly
  # ...
end

Info

Prior to julia 1.11, this constant was a global reference.

Auxiliary functions

StructPack.skip — Function

skip(io::IO)

Skip the msgpack value in io.

StructPack.step — Function

step(io::IO)

Enter or skip the msgpack value in io.

If the value is in map or vector format, step into it. Otherwise, skip the element.

Returns the core format of the value skipped or stepped into.

For example, if a msgpack array is stored in io, then step(io); unpack(io) will (generically) unpack the first element of the array.